LRCI-3544 Convert Git Branches to be extend new base Entity classes
