+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-oauth-application-headless-server/src/main/java/com/liferay/jethr0/project/dalo/ProjectsToGitBranchesDALO.java
@@ -1,0 +1,55 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.jethr0.project.dalo;
+
+import com.liferay.jethr0.entity.dalo.BaseEntityRelationshipDALO;
+import com.liferay.jethr0.entity.factory.EntityFactory;
+import com.liferay.jethr0.gitbranch.GitBranch;
+import com.liferay.jethr0.gitbranch.GitBranchFactory;
+import com.liferay.jethr0.project.Project;
+import com.liferay.jethr0.project.ProjectFactory;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * @author Michael Hashimoto
+ */
+@Configuration
+public class ProjectsToGitBranchesDALO
+	extends BaseEntityRelationshipDALO<Project, GitBranch> {
+
+	@Override
+	public EntityFactory<GitBranch> getChildEntityFactory() {
+		return _gitBranchFactory;
+	}
+
+	@Override
+	public EntityFactory<Project> getParentEntityFactory() {
+		return _projectFactory;
+	}
+
+	@Override
+	protected String getObjectRelationshipName() {
+		return "projectsToGitBranches";
+	}
+
+	@Autowired
+	private GitBranchFactory _gitBranchFactory;
+
+	@Autowired
+	private ProjectFactory _projectFactory;
+
+}
